redis multiple-nodes architecture (#85)
* fix #84

* update for #68: add node architecture

* add \n

* handle add or delete nodes ops

* handle change vxnet

* update log and checklist

* upgrade for comments

* upgrade for comments

* dfelay to delay

+++ ansible/roles/redis-standalone/files/opt/app/bin/node/redis-standalone.sh
@@ -2,6 +2,7 @@ FIND_MASTER_IP_ERR=210
 LEAVING_REDIS_NODES_IS_NONE_ERR=211
 LEAVING_REDIS_NODES_INCLUDE_MASTER_ERR=212
 MASTER_SALVE_SWITCH_WHEN_DEL_NODE_ERR=213
+LEAVING_REDIS_NODES_INCLUDE_SENTINEL_NODE=214
 REDIS_COMMAND_EXECUTE_FAIL_ERR=220
 RDB_FILE_EXIST_ERR=221
 RDB_FILE_CHECK_ERR=222
@@ -9,6 +10,12 @@ CLUSTER_STATUS_ERR=223
 NODES_NUMS_ERR=224
 CONFIRM_ERR=225
 BEYOND_DATABASES_ERR=226
+CLUSTER_STATS_ERR=227
+SENTINEL_RESET_ERR=228
+SENTINEL_FLUSH_CONFIG_ERR=229
+SENTINEL_START_ERR=230
+REDIS_START_ERR=231
+REDIS_STOP_ERR=232
 
 
 initNode() {
@@ -29,7 +36,7 @@ start() {
 stop() {
   if [ -z "$LEAVING_REDIS_NODES" ] && isSvcEnabled redis-sentinel; then
     stopSvc redis-sentinel
-    local sentinelHost; for sentinelHost in $REDIS_NODES; do
+    local sentinelHost; for sentinelHost in $STABLE_REDIS_NODES; do
       if [[ "${sentinelHost##*/}" != "$MY_IP" ]]; then
         retry 150 0.1 0 checkSentinelStopped ${sentinelHost##*/} || log "WARN: sentinel '$sentinelHost' is still up."
       fi
@@ -112,9 +119,37 @@ checkMasterNotLeaving() {
   if [[ "$LEAVING_REDIS_NODES " == *"/$master "* ]]; then return $LEAVING_REDIS_NODES_INCLUDE_MASTER_ERR; fi
 }
 
+findStableNodesCount() {
+  echo "$STABLE_REDIS_NODES" |wc -w
+}
+
 preScaleIn() {
   [ -n "$LEAVING_REDIS_NODES" ] || return $LEAVING_REDIS_NODES_IS_NONE_ERR
   checkMasterNotLeaving
+
+  # 防止最终计划剩余3个节点的时候，删除了 sentinel 节点
+  if [[ $(findStableNodesCount) -ge 3 ]];then
+    log --debug "check wether contain sentinel nodes"
+    local firstToThirdNode="$(getSentinelNodes)"
+    log "LEAVING_REDIS_NODES: $LEAVING_REDIS_NODES"
+    log "sentinel nodes: $firstToThirdNode"
+    local node;for node in $firstToThirdNode;do
+      echo "$LEAVING_REDIS_NODES" |grep -vq "$node" || {
+        log "leaving nodes include sentinel node: $node"
+        return $LEAVING_REDIS_NODES_INCLUDE_SENTINEL_NODE
+      }
+    done
+  fi
+}
+
+getSentinelNodes(){
+  eval echo "$STABLE_REDIS_NODES $LEAVING_REDIS_NODES" |xargs -n1| sort -n -t'/' -k1| xargs|cut -d" " -f1-3
+}
+
+getStableNodesIps(){
+  log --debug "STABLE_REDIS_NODES: $STABLE_REDIS_NODES"
+  log --debug "LEAVING_REDIS_NODES: $LEAVING_REDIS_NODES"
+  echo "$STABLE_REDIS_NODES" |xargs -n1 |awk -F/ '{print $3}'
 }
 
 destroy() {
@@ -122,11 +157,61 @@ destroy() {
     checkMasterNotLeaving
     execute stop
     checkVip || ( execute start && return $MASTER_SALVE_SWITCH_WHEN_DEL_NODE_ERR )
+    local stableNodesCount=$(findStableNodesCount)
+    log "stableNodesCount: $stableNodesCount"
+    [[ $stableNodesCount -ne 1 ]] || {
+      log "skip reset sentinel"
+      return 0
+      }
+    # determine redis-server is down for leaving nodes
+    log --debug "LEAVING_REDIS_NODES: $LEAVING_REDIS_NODES
+    MY_IP：$MY_IP
+    return code: $([[ "$LEAVING_REDIS_NODES " == *"/$MY_IP "* ]];echo $?)
+    "
+    if [[ "${LEAVING_REDIS_NODES%% *} " == *"/$MY_IP "* ]]; then
+      log --debug "start check redis-server down"
+      local leavingNode; for leavingNode in $LEAVING_REDIS_NODES; do
+        log "${leavingNode##*/} is down?"
+        retry 150 0.1 0 checkRedisStopped ${leavingNode##*/} || {
+          log "WARN: redis-server '$leavingNode' is still up."
+          return $REDIS_STOP_ERR
+          }
+        log "${leavingNode##*/} is down"
+      done
+      # 检查剩余的节点服务状态是 ok 的
+      log --debug "REDIS_NODES: $STABLE_REDIS_NODES"
+      log --debug "LEAVING_REDIS_NODES: $LEAVING_REDIS_NODES"
+      local stableNodesIps masterIp; stableNodesIps="$(getStableNodesIps)"
+      log "stableNodesIps: $stableNodesIps"
+      masterIp=$(findMasterIpByNodeIp $(echo "$stableNodesIps" |head -n1))
+      log --debug "masterIp: $masterIp"
+      local replicaResultOnline="$(runRedisCmd -h $masterIp info replication |grep "state=online")"
+      log --debug "replicaResultOnline: $replicaResultOnline"
+      local stableNodeIp; for stableNodeIp in $stableNodesIps; do
+        log "$stableNodeIp is ok?"
+        [[ "$stableNodeIp" == "$masterIp" ]] || echo "$replicaResultOnline" |grep -q "ip=${stableNodeIp//\./\\.}," || return $CLUSTER_STATS_ERR
+        log "$stableNodeIp is ok"
+      done
+      # 对所有的 sentinel 节点执行 sentinel reset
+      log "start sentinel reset"
+      local sentinelNodes; sentinelNodes="$(getSentinelNodes)"
+      log --debug "sentinelNodes: $sentinelNodes"
+      local sentinelNode; for sentinelNode in $sentinelNodes; do
+        log "${sentinelNode##*/} reset?"
+        runRedisCmd -h ${sentinelNode##*/} -p $SENTINEL_PORT SENTINEL RESET $SENTINEL_MONITOR_CLUSTER_NAME || return $SENTINEL_RESET_ERR
+        runRedisCmd -h ${sentinelNode##*/} -p $SENTINEL_PORT SENTINEL FLUSHCONFIG || return $SENTINEL_FLUSH_CONFIG_ERR
+        log "${sentinelNode##*/} reset"
+      done
+
+    fi
+
   fi
 }
 
 scaleIn() {
-  stopSvc redis-sentinel && rm -f $runtimeSentinelFile*
+  if [[ $(findStableNodesCount) -eq 1 ]]; then
+    stopSvc redis-sentinel && rm -f $runtimeSentinelFile*
+  fi 
 }
 
 restore() {
@@ -172,29 +257,70 @@ getRuntimeNameOfCmd() {
 
 runtimeSentinelFile=/data/redis/sentinel.conf
 
+getmasterIpFromSentinelNodes(){
+  log "getmasterIpFromSentinelNodes"
+  local sentinelNodes firstSentinelNode; sentinelNodes="$(getSentinelNodes)"
+  log --debug "sentinelNodes: $sentinelNodes"
+  local sentinelNode; for sentinelNode in $sentinelNodes; do
+    log "check ${sentinelNode##*/}?"
+    retry 60 0.5 0 checkSentinelStarted ${sentinelNode##*/} || {
+      log "${sentinelNode##*/} sentinel is not started"
+      return $SENTINEL_START_ERR
+    }
+    retry 60 0.5 0 checkRedisStarted ${sentinelNode##*/} || {
+      log "${sentinelNode##*/} redis-server is not started"
+      return $REDIS_START_ERR
+    }
+    log "check ${sentinelNode##*/} end"
+  done
+  firstSentinelNode="$(echo "$sentinelNodes" |cut -d" " -f1)"
+  log --debug "firstSentinelNode: $firstSentinelNode"
+  log --debug "result: $(runRedisCmd --ip ${firstSentinelNode##*/} -p $SENTINEL_PORT sentinel get-master-addr-by-name $SENTINEL_MONITOR_CLUSTER_NAME)"
+  local rc=0
+  runRedisCmd --ip ${firstSentinelNode##*/} -p $SENTINEL_PORT sentinel get-master-addr-by-name $SENTINEL_MONITOR_CLUSTER_NAME |xargs \
+        |awk '{if ($2 == '$REDIS_PORT') {print $1} else {'rc'=1;exit 1}}' || \
+        log "get master ip from ${firstSentinelNode##*/} fail! rc=$rc"
+  return $rc
+}
+
 getInitMasterIp() {
-  local firstRedisNode=${REDIS_NODES%% *}
+  local firstRedisNode=${STABLE_REDIS_NODES%% *}
   echo -n ${firstRedisNode##*/}
 }
 
 # 防止 revive 时从本地文件获取，导致双主以及vip 解绑
 getMasterIpForRevive() {
-  local rc=0
-  if isSvcEnabled redis-sentinel;then
-    local otherFirstNodeIp="$(echo $REDIS_NODES |awk 'BEGIN{RS=" "} {if ($1!~/'$MY_IP'$/) {print $1;exit 0}}'|awk 'BEGIN{FS="/"} {print $3}')"
-    runRedisCmd --ip ${otherFirstNodeIp} -p 26379 sentinel get-master-addr-by-name master |xargs \
-      |awk '{if ($2 == '$REDIS_PORT') {print $1} else {'rc'=1;exit 1}}' || \
-        log "get master ip from ${otherFirstNodeIp} fail! rc=$rc"
-    return $rc
+  local sentinelNodes="$(getSentinelNodes)"
+  log --debug "sentinelNodes: $sentinelNodes"
+  log --debug "MY_IP: $MY_IP"
+  if [[ "$sentinelNodes " == *"/$MY_IP "* ]]; then
+    local rc=0
+    if isSvcEnabled redis-sentinel;then
+      local otherFirstNodeIp="$(echo $STABLE_REDIS_NODES |awk 'BEGIN{RS=" "} {if ($1!~/'$MY_IP'$/) {print $1;exit 0}}'|awk 'BEGIN{FS="/"} {print $3}')"
+      runRedisCmd --ip ${otherFirstNodeIp} -p $SENTINEL_PORT sentinel get-master-addr-by-name $SENTINEL_MONITOR_CLUSTER_NAME |xargs \
+        |awk '{if ($2 == '$REDIS_PORT') {print $1} else {'rc'=1;exit 1}}' || \
+          log "get master ip from ${otherFirstNodeIp} fail! rc=$rc"
+      return $rc
+    else
+      log "get masterIp from init"
+      getInitMasterIp
+    fi
   else
-    getInitMasterIp
+    getmasterIpFromSentinelNodes
   fi
 }
 
 getMasterIpByConf() {
-  isSvcEnabled redis-sentinel && [ -f "$runtimeSentinelFile" ] \
-    && awk 'BEGIN {rc=1} $0~/^sentinel monitor master / {print $4; rc=0} END {exit rc}' $runtimeSentinelFile \
+  local sentinelNodes="$(getSentinelNodes)"
+  log --debug "sentinelNodes: $sentinelNodes"
+  log --debug "MY_IP: $MY_IP"
+  if [[ "$sentinelNodes " == *"/$MY_IP "* ]]; then
+    isSvcEnabled redis-sentinel && [ -f "$runtimeSentinelFile" ] \
+      && awk 'BEGIN {rc=1} $0~/^sentinel monitor '$SENTINEL_MONITOR_CLUSTER_NAME' / {print $4; rc=0} END {exit rc}' $runtimeSentinelFile \
       || getInitMasterIp
+  else
+    getmasterIpFromSentinelNodes
+  fi || getInitMasterIp
 }
 
 findMasterIp() {
@@ -221,7 +347,7 @@ backup(){
 }
 
 findMasterNodeId() {
-  echo $REDIS_NODES | xargs -n1 | awk -F/ '$3=="'$(findMasterIp)'" {print $2}'
+  echo $STABLE_REDIS_NODES | xargs -n1 | awk -F/ '$3=="'$(findMasterIp)'" {print $2}'
 }
 
 runRedisCmd() {
@@ -235,7 +361,7 @@ runRedisCmd() {
     echo "$result"
   fi
   return $retCode
-}
+} 
 
 checkVip() {
   local vipResponse; vipResponse="$(runRedisCmd --ip $REDIS_VIP ROLE | sed -n '1{p;q}')"
@@ -269,12 +395,20 @@ unbindVip() {
   ip addr del $REDIS_VIP/24 dev eth0 || [ "$?" -eq 2 ] # 2: not bound
 }
 
+checkSentinelStarted(){
+  nc -z -w3 $1 $SENTINEL_PORT
+}
+
 checkSentinelStopped() {
-  ! nc -z -w3 $1 26379
+  ! checkSentinelStarted
+}
+
+checkRedisStarted(){
+  nc -z -w3 $1 $REDIS_PORT
 }
 
 checkRedisStopped() {
-  ! nc -z -w3 $1 $REDIS_PORT
+  ! checkRedisStarted
 }
 
 encodeCmd() {
@@ -291,7 +425,10 @@ reload() {
   if [ "$1" == "redis-server" ]; then
     if [ -n "${JOINING_REDIS_NODES}" ]; then
       log --debug "scaling out ..."
-      configure && startSvc redis-sentinel
+      # 仅在从单个节点增加的时候才启动 sentinel
+      if [[ $(findStableNodesCount) -eq 1 ]]; then
+        configure && startSvc redis-sentinel
+      fi
     elif [ -n "${LEAVING_REDIS_NODES}" ]; then
       log --debug "scaling in ..."
     elif checkFileChanged $changedConfigFile; then
@@ -326,7 +463,7 @@ configureForChangeVxnet() {
   log --debug "exec configureForChangeVxnet"
   local runtimeConfigFile=/data/redis/redis.conf 
   sudo -u redis touch $nodesFile && rotate $nodesFile
-  echo $REDIS_NODES | xargs -n1 > $nodesFile
+  echo "$STABLE_REDIS_NODES" | xargs -n1 > $nodesFile
 
   if checkFileChanged $nodesFile; then
     log "IP addresses changed from [$(paste -s $nodesFile.1)] to [$(paste -s $nodesFile)]. Updating config files accordingly ..."
@@ -342,12 +479,14 @@ configureForSentinel() {
   local masterIp; masterIp="$(findMasterIp)"
   log --debug "masterIp is $masterIp"
   # flush every time even no master IP switches, but port is changed
-  echo "sentinel monitor master $masterIp $REDIS_PORT 2" > $monitorFile
+  echo "sentinel monitor $SENTINEL_MONITOR_CLUSTER_NAME $masterIp $REDIS_PORT 2" > $monitorFile
 
   if isSvcEnabled redis-sentinel; then
+    # 处理升级过程中监控名称的改变
+    sed -i 's/master/'"$SENTINEL_MONITOR_CLUSTER_NAME"'/g' $runtimeSentinelFile.1
     # 防止莫名的单个$0 出现
     awk '(NF>1 && $0~/^[^ #$]/) ? $1~/^sentinel/ ? $2~/^rename-/ ? !a[$1$2$3$4]++ : $2~/^(anno|deny-scr)/ ? !a[$1$2]++ : !a[$1$2$3]++ : !a[$1]++ : 0' \
-      $monitorFile $changedSentinelFile <(sed -r '/^sentinel (auth-pass master|rename-slaveof|rename-config|known-replica|known-slave)/d' $runtimeSentinelFile.1) > $runtimeSentinelFile
+      $monitorFile $changedSentinelFile <(sed -r '/^sentinel (auth-pass '"$SENTINEL_MONITOR_CLUSTER_NAME"'|rename-slaveof|rename-config|known-replica|known-slave)/d' $runtimeSentinelFile.1) > $runtimeSentinelFile
   else
     rm -f $runtimeSentinelFile*
   fi
@@ -397,7 +536,7 @@ restoreByCustomRdb(){
   uploadedRDBFile="/data/caddy/upload/dump.rdb" redisCheckRdbPath="/opt/redis/current/redis-check-rdb" destRDBfile="/data/redis/dump.rdb"
 
   # 仅允许单个节点做恢复数据操作
-  [[ $(echo "$REDIS_NODES" |xargs -n1 |wc -l) == 1 ]] || return $NODES_NUMS_ERR
+  [[ $(echo "$STABLE_REDIS_NODES" |xargs -n1 |wc -l) == 1 ]] || return $NODES_NUMS_ERR
   # 检查 RDB 文件是否 OK
   local myRole; myRole="$(getMyRole)"
   [[ "$myRole" == "master" ]] || {
@@ -432,9 +571,16 @@ check(){
 }
 
 getRedisRoles(){ 
-  local node nodeIp myRole; for node in $REDIS_NODES; do
+  local stableNodesCount=$(findStableNodesCount)
+  local node nodeIp myRole allow_deletion counter=0; for node in $(echo "$STABLE_REDIS_NODES" |sort -n -t"/" -k1); do
+    counter=$(($counter+1))
     nodeIp="$(echo "$node" |cut -d"/" -f3)"
-    myRole="$(runRedisCmd --ip "$nodeIp" role | head -n1)"
-    echo "$nodeIp $myRole"
-  done | jq -Rc 'split(" ") | [ . ]' | jq -s add | jq -c '{"labels":["ip","role"],"data":.}'
+    myRole="$(runRedisCmd --ip "$nodeIp" role | head -n1 || echo "unknown")"
+    if [[ $stableNodesCount -gt 3 ]]; then
+      if [[ $counter -le 3 ]];then allow_deletion="false";else allow_deletion="true";fi
+    else
+      if [[ "$myRole" == "master" ]]; then allow_deletion="false";else allow_deletion="true";fi
+    fi
+    echo "$nodeIp $myRole $allow_deletion"
+  done | jq -Rc 'split(" ") | [ . ]' | jq -s add | jq -c '{"labels":["ip","role","allow_deletion"],"data":.}'
 }
